* added more separate testcases
