+++ test/dependency-in-meta/galaxy.yml
@@ -1,0 +1,5 @@
+---
+
+dependencies:
+  # from galaxy
+  - src: yatesr.timezone
