+++ test/dependency-in-meta/gitlab.yml
@@ -1,0 +1,7 @@
+---
+
+dependencies:
+  # from GitLab or other git-based scm
+  - src: git@gitlab.company.com:mygroup/ansible-base.git
+    scm: git
+    version: "0.1"  # quoted, so YAML doesn't parse this as a floating-point value
